backend/requirements.txt

https://github.com/PnX-SI/Nomenclature-api-module/archive/1.3.0.zip
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.3
python-dateutil==2.8.0
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6